`clock_gettime` Don't multiply ns by resolution
Due to a poor reading of the scriptures, I misinterpreted the purpose of
`clock_getres`. This was not a problem on Linux amd64 where resolution
is commonly 1ns, but problematic in any system that has a resolution !=
1ns. The fix is to simply add seconds + (ns / 1e9).

A nanosecond is always a nanosecond; `clock_getres` simply provides the
expected sampling error for the given clock.

Fixes #8
Thanks to Russell Haley for the report.